+++ src/test/java/io/xpring/xrpl/XRPClientIntegrationTests.java
@@ -84,4 +84,17 @@ public class XRPClientIntegrationTests {
     List<XRPTransaction> paymentHistory = xrpClient.paymentHistory(XRPL_ADDRESS);
     assertThat(paymentHistory.size()).isGreaterThan(0);
   }
+
+  @Test
+  public void getPaymentTest() throws XRPException {
+    // GIVEN a hash of a payment transaction.
+    Wallet wallet = new Wallet(WALLET_SEED);
+    String transactionHash = xrpClient.send(AMOUNT, XRPL_ADDRESS, wallet);
+
+    // WHEN the transaction is requested.
+    XRPTransaction transaction = xrpClient.getPayment(transactionHash);
+
+    // THEN it is found and returned.
+    assertThat(transaction).isNotNull();
+  }
 }