+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- A new method, `getPayment`, added to `XRPClient` for retrieving payment transactions by hash.
 
 ## 5.1.1 - 2020-05-15